# Use fsspec.parquet for reading remote Parquet files