--- a/tests/nested_pandas/nestedframe/test_io.py
+++ b/tests/nested_pandas/nestedframe/test_io.py
@@ -399,3 +399,52 @@ def test__transform_read_parquet_data_arg():
                 "https://data.lsdb.io/hats/gaia_dr3/gaia/dataset/Norder=2/Dir=0/Npix=0.parquet",
             ]
         )
+
+
+def test_read_parquet_with_fsspec_optimization():
+    """Test that read_parquet automatically uses fsspec optimization for remote files."""
+    # Test with local file (should not use fsspec optimization)
+    local_path = "tests/test_data/nested.parquet"
+
+    # Test basic reading - local files should work as before
+    nf1 = read_parquet(local_path)
+
+    # Test with additional kwargs
+    nf2 = read_parquet(local_path, columns=["a", "nested.flux"], use_threads=True)
+
+    assert len(nf2) <= len(nf1)  # filtered columns
+    assert "a" in nf2.columns
+    assert "nested" in nf2.columns
+
+
+def test_fsspec_optimization_path_detection():
+    """Test _should_use_fsspec_optimization correctly identifies remote paths."""
+    from pathlib import Path
+
+    from nested_pandas.nestedframe.io import _should_use_fsspec_optimization
+
+    # Test cases that should NOT use optimization
+    assert not _should_use_fsspec_optimization("local_file.parquet", None)
+    assert not _should_use_fsspec_optimization("/path/to/file.parquet", None)
+    assert not _should_use_fsspec_optimization(Path("local_file.parquet"), None)
+    assert not _should_use_fsspec_optimization(UPath("local_file.parquet"), None)
+    assert not _should_use_fsspec_optimization("https://example.com/file.parquet", "some_filesystem")
+
+    # Test file-like object
+    import io
+
+    assert not _should_use_fsspec_optimization(io.BytesIO(b"test"), None)
+
+    # Test cases that SHOULD use optimization
+    assert _should_use_fsspec_optimization("https://example.com/file.parquet", None)
+    assert _should_use_fsspec_optimization("s3://bucket/file.parquet", None)
+    assert _should_use_fsspec_optimization("gs://bucket/file.parquet", None)
+    assert _should_use_fsspec_optimization(UPath("https://example.com/file.parquet"), None)
+    assert _should_use_fsspec_optimization(UPath("s3://bucket/file.parquet"), None)
+
+
+def test_docstring_includes_fsspec_notes():
+    """Test that the docstring mentions the automatic fsspec optimization."""
+    docstring = read_parquet.__doc__
+    assert "fsspec" in docstring
+    assert "remote" in docstring.lower()